app/requirements.txt
Flask
llama-cpp-python
